refactor(depot-client): split embedded depot transport
